CDVD: Purge AsyncFileReader interface
Everything goes through ThreadedFileReader now.
