+++ pcsx2/CDVD/CompressedFileReader.h
@@ -4,6 +4,8 @@
 #pragma once
 #include <string>
 
+class ThreadedFileReader;
+
 // Factory - creates an AsyncFileReader derived instance which can read a compressed file
 class CompressedFileReader
 {
@@ -12,7 +14,7 @@ public:
 	// If no matching handler is found, NULL is returned.
 	// The returned instance still needs ->Open(filename) before usage.
 	// Open(filename) may still fail.
-	static AsyncFileReader* GetNewReader(const std::string& fileName);
+	static ThreadedFileReader* GetNewReader(const std::string& fileName);
 
 private:
 	virtual ~CompressedFileReader() = 0;
