+++ pcsx2/CDVD/ThreadedFileReader.h
@@ -1,21 +1,28 @@
-// SPDX-FileCopyrightText: 2002-2023 PCSX2 Dev Team
+// SPDX-FileCopyrightText: 2002-2024 PCSX2 Dev Team
 // SPDX-License-Identifier: LGPL-3.0+
 
 #pragma once
 
-#include "AsyncFileReader.h"
+#include "common/Pcsx2Defs.h"
 
 #include <thread>
 #include <mutex>
 #include <atomic>
 #include <condition_variable>
 
+class Error;
+
 /// A file reader for use with compressed formats
 /// Calls decompression code on a separate thread to make a synchronous decompression API async
-class ThreadedFileReader : public AsyncFileReader
+class ThreadedFileReader
 {
 	ThreadedFileReader(ThreadedFileReader&&) = delete;
 protected:
+	std::string m_filename;
+
+	u32 m_dataoffset = 0;
+	u32 m_blocksize = 2048;
+
 	struct Chunk
 	{
 		/// Negative block IDs indicate invalid blocks
@@ -39,7 +46,6 @@ protected:
 	virtual void Close2() = 0;
 
 	ThreadedFileReader();
-	~ThreadedFileReader();
 
 private:
 	int m_amtRead;
@@ -96,12 +102,19 @@ private:
 	bool TryCachedRead(void*& buffer, u64& offset, u32& size, const std::lock_guard<std::mutex>&);
 
 public:
-	bool Open(std::string filename, Error* error) final override;
-	int ReadSync(void* pBuffer, u32 sector, u32 count) final override;
-	void BeginRead(void* pBuffer, u32 sector, u32 count) final override;
-	int FinishRead() final override;
-	void CancelRead() final override;
-	void Close() final override;
-	void SetBlockSize(u32 bytes) final override;
-	void SetDataOffset(u32 bytes) final override;
+	virtual ~ThreadedFileReader();
+
+	const std::string& GetFilename() const { return m_filename; }
+	u32 GetBlockSize() const { return m_blocksize; }
+
+	virtual u32 GetBlockCount() const = 0;
+
+	bool Open(std::string filename, Error* error);
+	int ReadSync(void* pBuffer, u32 sector, u32 count);
+	void BeginRead(void* pBuffer, u32 sector, u32 count);
+	int FinishRead();
+	void CancelRead();
+	void Close();
+	void SetBlockSize(u32 bytes);
+	void SetDataOffset(u32 bytes);
 };
