+++ pcsx2/pcsx2.vcxproj
@@ -460,7 +460,6 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Achievements.h" />
-    <ClInclude Include="AsyncFileReader.h" />
     <ClInclude Include="CDVD\BlockdumpFileReader.h" />
     <ClInclude Include="CDVD\CDVDdiscReader.h" />
     <ClInclude Include="CDVD\CompressedFileReader.h" />
